+++ mex/common/settings.py
@@ -2,15 +2,16 @@ from pathlib import Path
 from typing import Any, Self, cast
 
 from pydantic import AnyUrl, Field, SecretStr, model_validator
+from pydantic import BaseModel as PydanticBaseModel
 from pydantic_core import Url
 from pydantic_settings import BaseSettings as PydanticBaseSettings
 from pydantic_settings import SettingsConfigDict
 from pydantic_settings.sources import ENV_FILE_SENTINEL, DotenvType, EnvSettingsSource
 
-from mex.common.context import SingletonStore
+from mex.common.context import SingleSingletonStore
 from mex.common.types import AssetsPath, IdentityProvider, Sink, WorkPath
 
-SETTINGS_STORE = SingletonStore["BaseSettings"]()
+SETTINGS_STORE = SingleSingletonStore["BaseSettings"]()
 
 
 class BaseSettings(PydanticBaseSettings):
@@ -34,6 +35,7 @@ class BaseSettings(PydanticBaseSettings):
         env_prefix="mex_",
         env_file=".env",
         env_file_encoding="utf-8",
+        env_nested_delimiter="__",
         extra="ignore",
         validate_default=True,
         validate_assignment=True,
@@ -206,33 +208,17 @@ class BaseSettings(PydanticBaseSettings):
     @model_validator(mode="after")
     def resolve_paths(self) -> Self:
         """Resolve AssetPath and WorkPath."""
-        for name in self.model_fields:
-            value = getattr(self, name)
+
+        def _resolve(model: PydanticBaseModel, _name: str) -> None:
+            value = getattr(model, _name)
             if isinstance(value, AssetsPath) and value.is_relative():
-                setattr(self, name, self.assets_dir.resolve() / value)
+                setattr(model, _name, self.assets_dir.resolve() / value)
             elif isinstance(value, WorkPath) and value.is_relative():
-                setattr(self, name, self.work_dir.resolve() / value)
+                setattr(model, _name, self.work_dir.resolve() / value)
+            elif isinstance(value, PydanticBaseModel):
+                for sub_model_field_name in value.model_fields:
+                    _resolve(value, sub_model_field_name)
+
+        for name in self.model_fields:
+            _resolve(self, name)
         return self
-
-    @model_validator(mode="after")
-    def sync_settings(self) -> Self:
-        """Sync updates to settings in the `BaseSettings` scope to other settings.
-
-        Caveat: Updating fields of one class does not automatically update other
-        classes. To update another class, call any settings `.get()` method.
-        """
-        # ensure the settings singled instance is stored
-        SETTINGS_STORE.push(self)
-        # collect the changes to fields in the `BaseSettings` scope
-        base_scope = {
-            field: value
-            for field, value in self.model_dump(exclude_unset=True).items()
-            if field in BaseSettings.model_fields
-        }
-        # iterate over all active setting instances and swap them out
-        for settings in SETTINGS_STORE:
-            # create a new setting instance with `BaseSettings` fields are overwritten
-            SETTINGS_STORE.push(
-                settings.model_construct(**{**settings.model_dump(), **base_scope})
-            )
-        return cast(Self, SETTINGS_STORE.load(type(self)))